Refactor: change package path of Role enum
